Added local test for GetRecipient request
